Add wedding0 Fix issues
Add possible wedding effect, see description on DevidedSouls
+ yapf
